split app strings
