+++ doordeck-sdk-swift/Aux/AppStrings.swift
@@ -43,40 +43,4 @@ struct AppStrings {
     static let timeWindowsuccess: String = NSLocalizedString("Timer success", comment: "")
     static let timeWindowFailed: String = NSLocalizedString("Timer failed", comment: "")
     static let other: String = NSLocalizedString("Other errors", comment: "")
-    
-    static func messageForLockProgress(_ progress: LockDevice.currentUnlockProgress) -> String {
-        switch progress {
-        case .lockInitilized:
-            return AppStrings.lockInitilized
-        case .lockConnecting:
-            return AppStrings.lockConnecting
-        case .lockConnected:
-            return AppStrings.lockConnected
-        case .lockDisconnected:
-            return AppStrings.lockDisconnected
-        case .lockOffline:
-            return AppStrings.lockOffline
-        case .lockUnlocked:
-            return AppStrings.lockUnlocked
-        case .unlockSuccess:
-            return AppStrings.unlockSuccess
-        case .unlockFail:
-            return AppStrings.unlockFail
-        case .gpsFailed:
-            return AppStrings.gpsFailed
-        case .gpssuccess:
-            return AppStrings.gpssuccess
-        case .gpsSearching:
-            return AppStrings.gpsSearching
-        case .gpsUnauthorized:
-            return AppStrings.gpsUnauthorized
-        case .timeWindowsuccess:
-            return AppStrings.timeWindowsuccess
-        case .timeWindowFailed:
-            return AppStrings.timeWindowFailed
-        case .other:
-            return AppStrings.other
-        }
-    }
-
 }
